refactor: change id return value for unary expr
